[Hot-fix] Rename module 'auxiliary_hooks' -> 'recording_forward_hooks' (#89)
