+++ mpa/modules/models/segmentors/class_incr_segmentor.py
@@ -5,7 +5,7 @@
 import functools
 import torch
 from mmseg.utils import get_root_logger
-from mpa.modules.hooks.auxiliary_hooks import FeatureVectorHook
+from mpa.modules.hooks.recording_forward_hooks import FeatureVectorHook
 from mpa.modules.utils.task_adapt import map_class_names
 from mmseg.models import SEGMENTORS
 from mmseg.models.segmentors.encoder_decoder import EncoderDecoder
