+++ mpa/seg/inferrer.py
@@ -10,7 +10,7 @@ from mmcv.parallel import MMDataParallel
 from mmcv.runner import load_checkpoint, wrap_fp16_model
 from mmseg.datasets import build_dataloader, build_dataset
 from mmseg.models import build_segmentor
-from mpa.modules.hooks.auxiliary_hooks import FeatureVectorHook
+from mpa.modules.hooks.recording_forward_hooks import FeatureVectorHook
 from mpa.registry import STAGES
 from mpa.seg.stage import SegStage
 from mpa.stage import Stage
